fix: reference_in_meters could be a float value

+++ utils/conversions.py
@@ -2,7 +2,7 @@
 
 def convert_pixel_distance_to_meters(
     pixel_distance: float,
-    reference_in_meters: int,
+    reference_in_meters: float,
     reference_in_pixels: int,
 ) -> float:
     return (
@@ -13,7 +13,7 @@ def convert_pixel_distance_to_meters(
 
 def convert_meters_to_pixel_distance(
     meters: float,
-    reference_in_meters: int,
+    reference_in_meters: float,
     reference_in_pixels: int,
 ) -> int:
     return int(
